requirements.txt
pandas
numpy
pyro-ppl
matplotlib
tqdm
seaborn
black
isort
tabulate
scipy
requests
